test(random): #791 keep jtcop happy about the new tests
jtcop forbids static fields in a test class and a test class without a
production counterpart, so the pipeline count is a local now and the four
checks of the generator moved next to the test that uses it.
